LegalSchield/requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.14.2
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
connexion[uvicorn] >= 0.29.0
connexion[flask]
azure-storage-blob==12.19.1
